Ignore coverage on types
