+++ pyproject.toml
@@ -139,6 +139,9 @@ exclude_lines = [
     "^ +\\.\\.\\.$",
     "-> ['\"]?NoReturn['\"]?:",
 ]
+exclude_files = [
+    "src/prometheus_async/types.py"
+]
 
 
 [tool.black]
