+++ src/prometheus_async/types.py
@@ -1,0 +1,39 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Awaitable, Callable, Protocol
+
+
+if TYPE_CHECKING:
+    from prometheus_async.aio.web import MetricsHTTPServer
+
+
+Deregisterer = Callable[[], Awaitable[None]]
+
+
+class ServiceDiscovery(Protocol):
+    async def register(
+        self, metrics_server: MetricsHTTPServer
+    ) -> Deregisterer | None:
+        ...
+
+
+class Observer(Protocol):
+    def observe(self, value: float) -> None:
+        ...
+
+
+class Incrementer(Protocol):
+    def inc(
+        self, amount: float = 1, exemplar: dict[str, str] | None = None
+    ) -> None:
+        ...
+
+
+class IncDecrementer(Protocol):
+    def inc(
+        self, amount: float = 1, exemplar: dict[str, str] | None = None
+    ) -> None:
+        ...
+
+    def dec(self, amount: float = 1) -> None:
+        ...
